Added military time, editing delete user button

+++ app/Controllers/UsersController.js
@@ -2,6 +2,7 @@ import { appState } from "../AppState.js"
 import { User } from "../Models/User.js";
 import { usersService } from "../Services/UsersService.js"
 import { getFormData } from "../Utils/FormHandler.js"
+import { Pop } from "../Utils/Pop.js";
 import { setHTML } from "../Utils/Writer.js"
 
 function _drawUser() {
@@ -26,8 +27,9 @@ export class UsersController {
     usersService.newUser(newUser)
   }
 
-  deleteUser() {
+  async deleteUser() {
     window.event.preventDefault()
-    usersService.deleteUser()
+    if (await Pop.confirm("Are you sure you want to delete your acocunt?", "All your todos will be lost.", "Yes, I'm sure", "warning"))
+      await usersService.deleteUser()
   }
 }
